requirements.txt
Flask==3.1.0
Flask-SSLify==0.1.5
numpy==2.2.2
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
pydicom==3.0.1
requests==2.32.3
streamlit==1.41.1
